requirements.txt
discord.py==1.2.3
pymysql==0.9.3
colorama==0.3.9